% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uoe_scale.R
\name{ed_col}
\alias{ed_col}
\title{A helper function which modifies a given UoE Colour Palette by ramping to specified length}
\usage{
ed_col(palette = "bright", n, alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'ed_palettes' list}

\item{n}{Number of colours you want in your palette, defaults to n colours in palette}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
A helper function which modifies a given UoE Colour Palette by ramping to specified length
}
\examples{
library(UoEColouR)
image(volcano, col = ed_col("bright", 20))
}
