% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uoe_scale.R
\name{ed_pal}
\alias{ed_pal}
\title{Creates a palette from the UoE Edinburgh branded colours}
\usage{
ed_pal(palette = "bright", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'ed_palettes' list}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
Creates a palette from the UoE Edinburgh branded colours
}
