% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed_colours.R
\docType{data}
\name{ed_palettes}
\alias{ed_palettes}
\title{University of Edinburgh Brand Colours}
\format{
An object of class \code{list} of length 10.
}
\usage{
ed_palettes
}
\description{
A collection of colour palettes sourced from the University of Edinburgh Brand Colour Palette
}
\examples{
# Generate a plot using the default Edinburgh Bright palette
library(ggplot2)
ggplot(aes(x = hwy, y = cty, colour = manufacturer ), data = mpg) +
  geom_point() +
  scale_colour_uoe()



}
\keyword{datasets}
