% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uoe_scale.R
\name{scale_colour_uoe}
\alias{scale_colour_uoe}
\title{Creates a University of Edinburgh colour scale for ggplot2}
\usage{
scale_colour_uoe(
  palette = "bright",
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Choose from 'ed_palettes' list}

\item{discrete}{whether to use a discrete colour palette}

\item{alpha}{transparency}

\item{reverse}{logical, Reverse the order of the colours?}

\item{...}{additional arguments to pass to scale_color_gradientn}
}
\description{
Creates a University of Edinburgh colour scale for ggplot2
}
\examples{
library(ggplot2)
library(UoEColouR)

ggplot(mpg) +
 geom_point(aes(x = hwy, y = cty, color = manufacturer)) +
 scale_color_uoe("bright")

}
