% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uoe_scale.R
\name{scale_fill_uoe}
\alias{scale_fill_uoe}
\title{Creates a University of Edinburgh fill scale for ggplot2}
\usage{
scale_fill_uoe(
  palette = "bright",
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Choose from 'ed_palettes' list}

\item{discrete}{whether to use a discrete colour palette}

\item{alpha}{The alpha transparency, a number in [0,1], see argument alpha in
\code{\link[grDevices]{hsv}}.}

\item{reverse}{If TRUE, the direction of the colours is reversed.}

\item{...}{additional arguments to pass to scale_color_gradientn}
}
\description{
Creates a University of Edinburgh fill scale for ggplot2
}
\examples{
library(ggplot2)
library(UoEColouR)

ggplot(aes(x = manufacturer, fill = manufacturer), data = mpg) +
  geom_bar() +
  scale_fill_uoe()
}
