% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed_colours.R
\name{uoe_colour}
\alias{uoe_colour}
\title{uoe_colour
Extracts hex codes from the list of UoE brand colours. Helper function for uoe_pal()}
\usage{
uoe_colour(...)
}
\arguments{
\item{...}{Character names of the UoE brand colours}
}
\description{
uoe_colour
Extracts hex codes from the list of UoE brand colours. Helper function for uoe_pal()
}
