% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uoe_pal.R
\name{uoe_pal}
\alias{uoe_pal}
\title{Return function to interpolate a UoE colour palette. Helper function for scale_uoe}
\usage{
uoe_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in ed_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a UoE colour palette. Helper function for scale_uoe
}
