% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uoe_show_palette.R
\name{uoe_show_palette}
\alias{uoe_show_palette}
\title{Show a given colour palette from UoEColouR}
\usage{
uoe_show_palette(palette)
}
\arguments{
\item{palette}{vector of character hex RGB values}
}
\description{
Given a character vector (hex RGB values), display palette in graphics window.
}
\examples{
library(UoEColouR)
nord_show_palette("bright-two-tone")
}
