% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed_colours.R
\docType{data}
\name{ed_colours}
\alias{ed_colours}
\title{University of Edinburgh corporate colours}
\format{
An object of class \code{character} of length 25.
}
\usage{
ed_colours
}
\description{
University of Edinburgh corporate colours
}
\keyword{datasets}
