% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed_palettes.R
\docType{data}
\name{ed_palettes}
\alias{ed_palettes}
\title{Extracts hex codes from the list of UoE brand colours. Helper function for uoe_pal()}
\format{
An object of class \code{list} of length 8.
}
\usage{
ed_palettes
}
\description{
Extracts hex codes from the list of UoE brand colours. Helper function for uoe_pal()
}
\keyword{datasets}
