% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uoe_scale.R
\name{scale_fill_uoe}
\alias{scale_fill_uoe}
\title{Fill scale constructor for UoE branding colours}
\usage{
scale_fill_uoe(palette = "bright", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in ed_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for UoE branding colours
}
