% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UoE_colour.R
\name{uoe_colour}
\alias{uoe_colour}
\title{Extracts hex codes from the list of UoE brand colours}
\usage{
uoe_colour(...)
}
\arguments{
\item{...}{Character names of the UoE brand colours}
}
\description{
Extracts hex codes from the list of UoE brand colours
}
