% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uoe_scale.R
\name{uoe_pal}
\alias{uoe_pal}
\title{UoEColouR palettes delivered to ggplot preferences}
\usage{
uoe_pal(palette = "core", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Choose a palette from the named palettes in the 'ed_palettes' list}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
UoEColouR palettes delivered to ggplot preferences
}
